Bind components to Redux auth state
By connecting the Navigation and Book components to the Redux auth state,
we can then conditionally render certain parts of the app. For example,
if a user is not logged in, they will not see the "Add Review" button.
Logging in then displays said button as well as updates the nav bars in
the header and footer, replacing "Sign In" and "Get Started" with a
"Logout" button.